restore Kafka message parts impl to outbound adapter
