+++ crates/soup_realtime/Cargo.toml
@@ -10,7 +10,7 @@ consumer = [
   "dep:kafka_util",
   "dep:rdkafka",
   "dep:tokio",
-  "macro_event_broker/kafka",
+  "macro_event_broker/outbound",
   "ports",
 ]
 default = ["consumer", "outbound", "ports"]
